Se modifica el JSON para carga rápida y se organizan archivos
Se crea un nuevo script de Python que modifica el JSON entregado por el otro script agregando un hijo mayor 'COORDS' para obtener rápidamente la posición de las exploraciones.

Se organizan los archivos en carpetas, esto puede causar que algunos recursos no carguen y se deba modificar el relative path (en los scripts de Python).
