+++ map.html
@@ -81,7 +81,7 @@
               </div>
               <div class="tab-pane fade" id="nav-structures" role="tabpanel" aria-labelledby="nav-home-tab">
                 <div style="max-width: 70%; height: auto; margin:auto; padding: 40px">
-                  <img src="porcentaje.png" class="img-fluid " alt="smaple image">
+                  <img src="/images/porcentaje.png" class="img-fluid " alt="smaple image">
                 </div>
               </div>
               <div class="tab-pane fade" id="nav-sondeos" role="tabpanel" aria-labelledby="nav-profile-tab"></div>
